refactored models to not include ThingToKnow model, rulelist is now on Listing object
